[PTCD-177] isolated FAOC config in separate folder
